Initial pass at adding definition caching.
Included a cache implementation which writes plugin definitions to disk as anonymous classes which can simply be "include"ed in to load the cached definitions.

Also added a plugin definition trait to the code base since this was ubiquitous but only existed in plugins_annotation as the base annotation implementation.